requirements.txt
# =============================================================================
# S-ACM - Requirements for Railway/Production Deployment
# =============================================================================

# Django Core
Django>=5.1,<6.1
gunicorn>=24.0
whitenoise>=6.5
python-dotenv>=1.0
dj-database-url>=2.0

# Database
psycopg2-binary>=2.9

# Storage (Supabase S3)
boto3>=1.34
django-storages>=1.14

# Authentication & Security
# (uses Django built-in)

# API & AI
openai>=1.0
Markdown>=3.5

# File Processing
pdfplumber>=0.10
python-docx>=1.0
python-pptx>=0.6
Pillow>=10.0
PyPDF2>=3.0

# Excel Export
openpyxl>=3.1

# Admin & Forms
# django-crispy-forms>=2.0
# crispy-bootstrap5>=2024.2

# Async (Optional - Celery)
# celery>=5.3
# redis>=5.0

# Utilities
requests>=2.31
